Angle of Incidence (AOI)

Angle of Incidence (AOI)

AOI(azimuth, zenith, array.tilt, array.azimuth)

Arguments

azimuth
the solar zenith angle

zenith
the solar azimuth angle

array.tilt
the PV tilt angle

array.azimuth
the PV azimuth angle

Value

Details

\[AOI = \arccos\big({\cos{(zenith)}\cos{(array.tilt)}+ \sin{(zenith)}\sin{(array.tilt)}\cos{(azimuth-array.azimuth)}\big)}\] Though the equation returns AOI values with any given set of zenith, azimuth, array.tilt, and array.azimuth, the AOI is meaningful only if \(0 \leq zenith \leq 90\), and \(0 \leq AOI \leq 90\). The former condition assures the sun is above the horizon and the latter condition assures the sunlight beam is able to hit the panel (Source: Stackhouse et al., 2018).