Angle of Incidence (AOI)

Angle of Incidence (AOI)

angle_of_incidence(
  x,
  array.type = "fh",
  suffix = TRUE,
  azimuth_Q = "azimuth_Q",
  zenith = "zenith",
  na.val = NA,
  zenith.max = 90,
  AOI.max = 90,
  verbose = getOption("merra2.verbose"),
  ...
)

fAOI(
  x,
  array.type = "fh",
  suffix = TRUE,
  azimuth_Q = "azimuth_Q",
  zenith = "zenith",
  na.val = NA,
  zenith.max = 90,
  AOI.max = 90,
  verbose = getOption("merra2.verbose"),
  ...
)

Arguments

x

array.type

suffix

azimuth_Q

the solar zenith angle, degrees

zenith

the solar azimuth angle, degrees

na.val

zenith.max

verbose

...

Details

\[AOI = \arccos\big({\cos{(zenith)}\cos{(array.tilt)}+ \sin{(zenith)}\sin{(array.tilt)}\cos{(azimuth-array.azimuth)}\big)}\] Though the equation returns AOI values with any given set of zenith, azimuth, array.tilt, and array.azimuth, the AOI is meaningful only if \(0 \leq zenith \leq 90\), and \(0 \leq AOI \leq 90\). The former condition assures the sun is above the horizon and the latter condition assures the sunlight beam is able to hit the panel (Source: Stackhouse et al., 2018).

Examples

NA
#> [1] NA