Title

Title

Usage

fWindCF(
  x,
  height = 50,
  mps = paste0("W", height, "M"),
  return_name = paste0("win", height, "af"),
  hellmann = "hellmann",
  W10M = "W10M",
  W50M = "W50M",
  WPC = fWindPowerCurve,
  verbose = TRUE,
  ...
)

Arguments

x

data frame with MERRA-2 subset

height

height over ground

mps

name of wind speed variable available in `x` or will be extrapolated using `fHellmann` and `fWSE` functions

return_name

name of the variable, which will be added (or overwritten) to `x`

hellmann

name of the variable with Hellmann constant, either available in `x` or will be calculated using `fHellmann` function

W10M

name of the variable with wind speed at 10 meters height

W50M

name of the variable with wind speed at 10 meters height

WPC

name of the wind power capacity function

verbose

if TRUE, the process will be reported

...

additional parameters for `fHellmann`, `fWSE`, `WPC` functions

Value

`x` with added (or overwritten) column of wind power capacity factors; the name of the column is given by `return_name` parameter.

Examples

NA
#> [1] NA