Get MERRA-2 grid IDs which overlaps with the given spatial object

Get MERRA-2 grid IDs which overlaps with the given spatial object

Usage

get_locid(
  x,
  method = "polygons",
  add_poles_points = F,
  return_grid = FALSE,
  ...
)

Arguments

x

spatial (`sf`) object. Expected CRS is `EPSG:4326`, if different, it will be attempted to transform the grid to CRS of `x`, though intersection algorithms may not work well for some projections. Use `sf::st_transform(x, "EPSG:4326")` and `sf::st_make_valid(x)` to transform and correct geometries.

method

character, "polygons" (default) or "points" grid to use in `sf::st_intersects` method.

...

ignored