Get MERRA-2 subset

Get MERRA-2 subset

get_merra2_subset(
  locid = 1:207936L,
  from = "1980-01-01 00",
  to = "2020-12-31 23",
  tz = "UTC",
  cols = NULL,
  quiet = FALSE,
  rows_lim = 2 * 10^9,
  as_integers = FALSE
)

Arguments

locid

integer vector with location locid (values from 1 to 207936)

from

starting date and time of the subset in "YYYY-MM-DD HH" format

to

ending date and time of the subset in "YYYY-MM-DD HH" format (see fDate)

tz

time zone of the requested date-time interval (see OlsonNames)

cols

names or indexes of columns from the database (NULL default, returning all available columns)

quiet

if process should be reported

rows_lim

original.units

Value

data.table with the subset of

References

<https://en.wikipedia.org/wiki/List_of_tz_database_time_zones>

Examples

NA
#> [1] NA