Photovoltaic Solar Panel Orientation and Performance Models

Photovoltaic Solar Panel Orientation and Performance Models

pv_array_position(
  x,
  array.type = "fl",
  suffix = TRUE,
  lat = "lat",
  azimuth_Q = "azimuth_Q",
  zenith = "zenith",
  verbose = getOption("merra2.verbose"),
  tilt.param = tilt.param.default()
)

Arguments

x

data.frame object with MERRA-2 subset

array.type

type of tracking ()

suffix

lat

latitude of PV location (\(-90 \leq lat \leq 90\))

azimuth_Q

solar azimuth angle for the PV location (\(0 \leq azimuth < 360\))

zenith

solar zenith angle for the PV location (\(0 \leq azimuth \leq 90\))

verbose

tilt.param

Details

Fixed PV Panel (\(*.fl\))

South-facing fixed solar PV with the tilted angle equal to the site's latitude. PV Tilted Angle (\(array.tilt\), in degrees) \[array.tilt = latitude\] PV Azimuth Angle (\(array.azimuth\), in degrees) \[array.azimuth = 180\]
Horizontal Single-Axis PV Tracker (\(*.th\))

A horizontal single-axis PV tracker with its axis in align with the meridian direction and parallel to the ground. PV tilted angle under the optimal rotation strategy (\(array.tilt\), in degrees) \[array.tilt = \arctan{\big(\tan{(zenith)}\cos{(azimuth-array.azimuth)}\big)}\] PV Azimuth Angle (\(array.azimuth\), in degrees) \[array.azimuth = 90 \textrm{ if } azimuth < 180\] \[array.azimuth = 270 \textrm{ if } azimuth \geq180\]
Vertical Single-Axis Tracker (\(*.tv\))

A vertical single-axis PV tracker with its axis normal to the ground. PV tilted angle under the optimal rotation strategy (\(array.tilt\), in degrees) \[array.tilt = latitude\] PV Azimuth Angle (\(array.azimuth\), in degrees) \[array.azimuth = azimuth\]
Tilted Single-Axis Tracker (\(*.tl\))

A single-axis PV tracker with its axis parallel to the meridian direction and the axis tilted angle equal to the site’s latitude. PV tilted angle under the optimal rotation strategy (\(array.tilt\), in degrees) \[array.tilt = \arctan{\big(\frac{\tan{(zenith)}} {\cos{(array.azimuth-180)}}\big)}+\delta\pi\] where \(\delta=0 \textrm{ when } 90 \leq array.azimuth \leq 270, \textrm{ otherwise }\delta=1\) PV Azimuth Angle (\(array.azimuth\), in degrees) \[array.azimuth = 180(1+\sigma)+\Delta tilt\] where \[\sigma = \begin{cases} 1& & {\Delta tilt < 0, azimuth \geq 180}\newline 0& & {\Delta tilt \times (azimuth-180) \geq 0}\newline -1& & {\Delta tilt > 0, azimuth < 180} \end{cases}\] \[\Delta tilt = \arctan{\frac{\sin{(zenith)} \sin{(azimuth-180)}}{\cos{(\beta)}\sin{(latitude)}}}\] \[\cos{(\beta)} = \cos{(zenith)}\cos{(latitude)}+ \sin{(zenith)}\sin{(latitude)}\cos{(azimuth-180)}\]
Dual-Axis Tracker (\(*.td\))

A dual-axis PV tracker. PV Tilted Angle under the Optimal Rotation Strategy (\(array.tilt\), in degrees) \[array.tilt = zenith\] PV Azimuth Angle under the Optimal Rotation Strategy (\(array.azimuth\), in degrees) \[array.azimuth = azimuth\]

Examples

NA
#> [1] NA