Wind power curve function. Estimates power output based on the speed of wind.

Wind power curve function. Estimates power output based on the speed of wind.

Usage

fWindPowerCurve(mps, cutin = 3, cutoff = 25, data = NULL, ...)

fWPC(mps, cutin = 3, cutoff = 25, data = NULL, ...)

Arguments

mps

wind speed, meters per second

cutin

minimal speed of production

cutoff

maximum speed of wind

data

data frame or list with two power curve data, columns `speed` and `af`

Value

estimated capacity factor of a wind turbine

Examples

fWPC(0:30)
#>  [1] 0.000 0.000 0.000 0.000 0.017 0.066 0.138 0.235 0.362 0.518 0.688 0.840
#> [13] 0.941 0.983 0.995 0.999 1.000 1.000 1.000 1.000 1.000 1.000 1.000 1.000
#> [25] 1.000 1.000 0.000 0.000 0.000 0.000 0.000
plot(0:35, WindPowerCurve(0:35), type = "l", col = "red", lwd = 2)
#> Error in WindPowerCurve(0:35): could not find function "WindPowerCurve"